+++ lib/i18n_yaml_editor/category.rb
@@ -2,7 +2,7 @@
 
 require 'set'
 
-module Translator
+module I18nYamlEditor
   # This is a category
   class Category
     attr_accessor :name, :keys