+++ lib/i18n_yaml_editor/store.rb
@@ -3,14 +3,14 @@
 require 'set'
 require 'pathname'
 
-require 'translator/transformation'
-require 'translator/filter'
-require 'translator/update'
-require 'translator/category'
-require 'translator/key'
-require 'translator/translation'
+require 'i18n_yaml_editor/transformation'
+require 'i18n_yaml_editor/filter'
+require 'i18n_yaml_editor/update'
+require 'i18n_yaml_editor/category'
+require 'i18n_yaml_editor/key'
+require 'i18n_yaml_editor/translation'
 
-module Translator
+module I18nYamlEditor
   # Raised when translation entries have the same key
   class DuplicateTranslationError < StandardError; end