+++ lib/i18n_yaml_editor/translation.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module Translator
+module I18nYamlEditor
   # A Translation holds information about its name, file and text
   class Translation
     attr_accessor :name, :file, :text