+++ test/test_helper.rb
@@ -11,13 +11,13 @@ require 'minitest/autorun'
 require 'minitest/hell'
 require 'minitest/reporters'
 require 'rack/test'
-require 'translator'
+require 'i18n_yaml_editor'
 
 Minitest::Reporters.use! [Minitest::Reporters::DefaultReporter.new(color: true)]
 
 module Minitest
   class Test
-    include Translator
+    include I18nYamlEditor
     parallelize_me!
   end
 end